Stabilize testSnapshotDeleteRelocatingPrimaryIndex (#61088)
Use transport blocking to make relocation take forever instead of relying on the relocation to take long enough to clash with the snapshot.

Closes #61069
